+++ semcore/wizard/src/Wizard.types.ts
@@ -95,7 +95,7 @@ export type WizardStepperProps<T extends WizardStep = WizardStep> = BoxProps & {
 
 export type WizardContentProps = BoxProps & {
   /**
-   * Render wizard content container with border-radius on the left side.
+   * Renders wizard content container with border-radius on the left side
    */
   noSidebar?: boolean;
 };